AquaDetect_web/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
torch==2.6.0
torchaudio==0.11.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.74
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
